saving: we can now write severals frames in parallel.
To control saving concurrency, you have to change the number of writing task
with method get/setMaxConcurrentWritingTask.

This commit contains also a small fix for write statistics:
As we recreate the container when changing saving type, thoses parameters
were not keeped.

We now set event callback in case of error when writing.